+++ Integrations/RSANetWitness_v11_1/RSANetWitness_v11_1.py
@@ -273,10 +273,13 @@ def get_incidents_request(since=None, until=None, page_number=None, page_size=10
     return response
 
 
-def get_all_incidents(since=None, until=None, limit=None):
+def get_all_incidents(since=None, until=None, limit=None, page_number=0):
     """
 
-    returns all/up to limit incidents in a time window
+    returns 
+    1. all/up to limit incidents in a time window
+    2. has_next
+    3. next_page
 
     """
 
@@ -284,7 +287,6 @@ def get_all_incidents(since=None, until=None, limit=None):
     if not limit:
         limit = float('inf')
     has_next = True
-    page_number = 0
     incidents = []  # type: list
     LOG('Requesting for incidents in timeframe of: {s} - {u}'.format(s=since or 'not specified',
                                                                      u=until or 'not specified'))
@@ -304,8 +306,47 @@ def get_all_incidents(since=None, until=None, limit=None):
 
     # if incidents list larger then limit - fit to limit
     if len(incidents) > limit:
-        incidents[limit - 1: -1] = []
+        incidents = incidents[:limit]
 
+    return incidents, has_next, page_number
+
+
+def get_all_incidents_from_beginning(since=None, until=None, limit=None, page_number=None, ignore_id=None):
+    """
+
+    returns
+    1. all/up to limit incidents in a time window
+    2. has_next
+    3. next_page
+
+    """
+    # if limit is None, set to infinity
+    if not limit:
+        limit = float('inf')
+    has_next = True
+    page_number = 0
+    incidents = []  # type: list
+    LOG('Requesting for incidents in timeframe of: {s} - {u}'.format(s=since or 'not specified',
+                                                                     u=until or 'not specified'))
+    while has_next:
+        # call get_incidents_request(), given user arguments
+        # returns the response body on success
+        # raises an exception on failed request
+        LOG('Requesting for page {}'.format(page_number))
+        response_body = get_incidents_request(
+            since=since,
+            until=until,
+            page_number=page_number
+        )
+        incidents.extend(response_body.get('items'))
+        has_next = response_body.get('hasNext')
+        page_number += 1
+
+    incidents.reverse()
+    incidents = list(filter(lambda inc: inc.get('id') != ignore_id, incidents))
+    # if incidents list larger then limit - fit to limit
+    if len(incidents) > limit:
+        return incidents[:limit]
     return incidents
 
 
@@ -336,11 +377,15 @@ def get_incidents():
     # parse limit argument to int
     if limit:
         limit = int(limit)
+    page_number = args.get('pageNumber')
+    if page_number:
+        page_number = int(page_number)
 
-    incidents = get_all_incidents(
+    incidents, has_next, next_page = get_all_incidents(
         since=args.get('since'),
         until=args.get('until'),
-        limit=limit
+        limit=limit,
+        page_number=page_number
     )
 
     md_content = create_incidents_list_md_table(incidents)
@@ -356,6 +401,9 @@ def get_incidents():
             "NetWitness.Incidents(obj.id==val.id)": incidents
         }
     }
+    if has_next:
+        entry['EntryContext']['NetWitness.Incidents(val.NextPage).NextPage'] = next_page
+        entry['HumanReadable'] += '\n## Not all incidents were fetched. Next page: {}'.format(next_page)
     demisto.results(entry)
 
 
@@ -604,15 +652,18 @@ def fetch_incidents():
     # if last timestamp was recorded- use it, else generate timestamp for one day prior to current date
     if last_run and last_run.get('timestamp'):
         timestamp = last_run.get('timestamp')
+        ignore_id = last_run.get('ignore_id')
     else:
         last_fetch, _ = parse_date_range(FETCH_TIME)
         # convert to ISO 8601 format and add Z suffix
         timestamp = last_fetch.isoformat() + 'Z'
+        ignore_id = None
 
     LOG('Fetching incidents since {}'.format(timestamp))
-    netwitness_incidents = get_all_incidents(
+    netwitness_incidents = get_all_incidents_from_beginning(
         since=timestamp,
-        limit=FETCH_LIMIT
+        limit=FETCH_LIMIT,
+        ignore_id=ignore_id
     )
 
     demisto_incidents = []
@@ -648,7 +699,11 @@ def fetch_incidents():
         demisto_incidents.append(parse_incident(incident))
 
     demisto.incidents(demisto_incidents)
-    demisto.setLastRun({'timestamp': last_incident_timestamp})
+    last_run = {'timestamp': last_incident_timestamp}
+    if netwitness_incidents:
+        last_run['ignore_id'] = netwitness_incidents[-1].get('id')
+    demisto.setLastRun(last_run)
+    return demisto_incidents
 
 
 def parse_incident(netwitness_incident):
@@ -950,7 +1005,7 @@ def test_module():
     since = datetime.now() - timedelta(days=int(10))
     timestamp = since.isoformat() + 'Z'
 
-    incidents = get_all_incidents(
+    incidents, _, __ = get_all_incidents(
         since=timestamp,
         until=None,
         limit=100